DATAREST-962 - Fixed imports in Gemfire integration tests after move to Geode.
